Move the list of alwratha into data-enty controller
In preperation to display singular relatives without input text
addon, we add field to each person in alwratha list indicating
that person singularity.

Because of the modification above, it appears that alwratha list
should be treated and thought of as part the model not the view.
Thus we move it into the controller as intermediate step before
eventually defining it in the model