Define navbar-brand as items[0] inside the repective controller
Previously, the mixin displayNavItems() used hard coded navbar
brand name. This resulted on all views showing 'the main screen'
as navbar-brand. However, the brand for home page should be the
'about' page.

To resolve this, we send the brand value as the first element of
items object and modify the logic inside the mixin to reflect
these design changes