Use property getter in mirath rules
Our first design was to use property functions to be executed when
the field is looked up. However, it turned out that property
function returns function object when its property is looked up.
To get a dynamically calculated value when looking up a property,
we use propery getters which get the function's returned value