Save & retrieve page state on back navigation
Before this, when user navigates to previous page, a fresh page is
served and hence all entered data would be lost. To improve UX in
editing the entered data, we save page state in sessionStorage and
retrieve it whenever the page is reloaded. The session data is
cleared at loading the home page of mirath.